requirements.txt
pandas<=2.0.2
python-decouple<=3.8
Requests<=2.31.0
